all: Add CN_470_510 band definition
