+++ pkg/band/all.go
@@ -35,6 +35,14 @@ var All = map[string]map[ttnpb.PHYVersion]Band{
 		ttnpb.RP001_V1_1_REV_A:   AU_915_928_RP1_v1_1_RevA,
 		ttnpb.RP001_V1_1_REV_B:   AU_915_928_RP1_v1_1_RevB,
 	},
+	CN_470_510: {
+		ttnpb.TS001_V1_0_1:       CN_470_510_TS1_v1_0_1,
+		ttnpb.RP001_V1_0_2:       CN_470_510_RP1_v1_0_2,
+		ttnpb.RP001_V1_0_2_REV_B: CN_470_510_RP1_v1_0_2_RevB,
+		ttnpb.RP001_V1_0_3_REV_A: CN_470_510_RP1_v1_0_3_RevA,
+		ttnpb.RP001_V1_1_REV_A:   CN_470_510_RP1_v1_1_RevA,
+		ttnpb.RP001_V1_1_REV_B:   CN_470_510_RP1_v1_1_RevB,
+	},
 }
 
 // Get returns the band if it was found, and returns an error otherwise.
